Add a default value for the category to fix test issues (#80)
The policy report tests have been behaving poorly.  In environments
where there is an insights NonCompliance, I think a metric gets created
that causes the tests to pass. In some environments that NonCompliance
doesn't exist causing this problem -- mainly because the policy resource
provided doesn't have a NIST 800-53 category defined so no metric is
created by the test. Providing a reasonable default for the category
makes sense to me since the standard/category/controls for organizing
compliance is optional -- but we don't want to optionally create the
metric.

Refs:
 - https://issues.redhat.com/browse/ACM-8851

Signed-off-by: Gus Parvin <gparvin@redhat.com>
Co-authored-by: Gus Parvin <gparvin@redhat.com>