Deduplicate metrics from policy report results (#148)
Because GRC policy results are named based on the Policy, but created
for each template, there can be multiple entries in the PolicyReport
results for what looks like the 'same' policy. If all of the metric
labels would be the same for any of those entries (eg, excluding the
description field), then a `PrometheusDuplicateTimestamps` might be
fired.

Now, duplicate results in a PolicyReport will increment the value of the
metric, instead of being incorrectly reported multiple times. This
handles not only the GRC case, but any source that puts results in the
PolicyReport.

Refs:
 - https://issues.redhat.com/browse/ACM-16014

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>